# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Task Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/tasks:
    get:
      operationId: listProjectTasks
      tags: [tasks]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task (Admin only)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (Admins can update anything; assignees can update status)
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task (Admin only)
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        imageUrl: { type: string, nullable: true }
      required: [id, email, name]
...
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
    TaskPriority:
      type: string
      enum: [low, medium, high]
...
    Task:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assigneeId: { type: string, nullable: true }
        assignee:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        dueDate: { type: string, format: date-time, nullable: true }
        createdById: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - projectId
        - title
        - status
        - priority
        - createdById
        - createdAt
        - updatedAt
...
    CreateTaskInput:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 140 }
        description: { type: string, maxLength: 2000, nullable: true }
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assigneeId: { type: string, nullable: true }
        dueDate: { type: string, format: date-time, nullable: true }
      required: [title, status, priority]
    UpdateTaskInput:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 140 }
        description: { type: string, maxLength: 2000, nullable: true }
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assigneeId: { type: string, nullable: true }
        dueDate: { type: string, format: date-time, nullable: true }
...