# lib/api-zod/src/generated/types/updateTaskInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from "./taskPriority";
import type { TaskStatus } from "./taskStatus";

export interface UpdateTaskInput {
  /**
   * @minLength 1
   * @maxLength 140
   */
  title?: string;
  /** @maxLength 2000 */
  description?: string | null;
  status?: TaskStatus;
  priority?: TaskPriority;
  assigneeId?: string | null;
  dueDate?: Date | null;
}